backend_ml/requirements.txt
# EquiTable Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pydantic>=2.5.3
pydantic-settings>=2.1.0
crawl4ai>=0.8.0
firecrawl-py>=0.0.16
google-generativeai>=0.8.0
python-dotenv>=1.0.0
httpx>=0.26.0
google-genai

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
